# docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: TekClinic REST API
  description: API of TekClinic
  version: 1.0.0
servers:
- url: /
paths:
  /patients:
    get:
      tags:
      - Patient
      description: returns all registered patients
      operationId: getPatients
      parameters:
      - name: skip
        in: query
        description: number of patients to skip
        required: false
        style: form
        explode: true
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        description: max number of patients to return
        required: false
        style: form
        explode: true
        schema:
          maximum: 50
          minimum: 1
          type: integer
          format: int32
          default: 20
      - name: search
        in: query
        description: Search term for filtering. It searches personal ID, phone number, name, special note, and referred person. All words in the search term must appear in one of the fields as a prefix.
        required: false
        style: form
        explode: true
        schema:
          maxLength: 100
          minLength: 1
          type: string
      responses:
        "200":
          description: all the patients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NamedAPIResourceList'
              example:
                count: 248
                next: http://api.tekclinic.org/patients/?limit=20&offset=20&search=john
                results:
                - id: 1
                  name: patient
                  url: http://api.tekclinic.org/patients/1/
                - id: 2
                  name: patient
                  url: http://api.tekclinic.org/patients/2/
    post:
      tags:
      - Patient
      operationId: createPatient
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientBase'
      responses:
        "201":
          description: id of the created patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
  /patients/{id}:
    get:
      tags:
      - Patient
      description: returns a specific patient
      operationId: getPatient
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: patient associated with the id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
    put:
      tags:
        - Patient
      description: updates a specific patient
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          description: id of the patient to be updated
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        "200":
          description: id of the updated patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdHolder'
    delete:
      tags:
        - Patient
      description: deletes a patient
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: patient deleted successfully
...
components:
  schemas:
    NamedAPIResourceList:
      required:
      - count
      - next
      - previous
      - results
      type: object
      properties:
        count:
          type: number
          description: total number of resources available from this API.
          format: int32
        next:
          type: string
          description: URL of the next page in the list.
          format: uri
          nullable: true
        previous:
          type: string
          description: URL of the previous page in the list.
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/NamedAPIResource'
    NamedAPIResource:
      required:
      - id
      - name
      - url
      type: object
      properties:
        id:
          type: integer
          description: identifier of the referenced resource.
          format: int32
        name:
          type: string
          description: type name of the referenced resource.
        url:
          type: string
          description: URL of the referenced resource.
          format: uri
          deprecated: true
    PatientBase:
      required:
      - birth_date
      - name
      - personal_id
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: full name of patient.
          example: Henry Smith
        personal_id:
          $ref: '#/components/schemas/PersonalId'
        gender:
          type: string
          description: gender of patient.
          example: male
          default: unspecified
          enum:
          - unspecified
          - male
          - female
        phone_number:
          type: string
          description: "phone number of patient, that includes country code."
          example: +972505201591
        languages:
          maxItems: 10
          type: array
          description: preferred languages of the patient.
          example:
          - Hebrew
          - Spanish
          items:
            type: string
          default: []
        birth_date:
          type: string
          description: birth date of patient.
          format: date
          example: 2000-07-21
        emergency_contacts:
          maxItems: 10
          type: array
          description: phone numbers of emergency contacts.
          items:
            $ref: '#/components/schemas/EmergencyContact'
          default: []
        referred_by:
          maxLength: 100
          type: string
          description: by who the patient got to the clinique.
          example: John
        special_note:
          maxLength: 500
          type: string
          description: additional notes.
          example: Can't answer calls between 12:00 and 20:00
    Patient:
      allOf:
      - required:
        - active
        - age
        - birth_date
        - emergency_contacts
        - gender
        - id
        - languages
        - name
        - personal_id
        type: object
        properties:
          id:
            type: integer
            description: identifier of the patient.
            format: int32
            example: 228
          active:
            type: boolean
            description: whether the patient still visits the clinic.
            example: true
            default: true
          age:
            type: integer
            description: estimated age of patient in years.
            format: int32
            example: 23
      - $ref: '#/components/schemas/PatientBase'
    PatientUpdate:
      allOf:
        - $ref: '#/components/schemas/PatientBase'
        - required:
            - active
            - age
            - name
            - personal_id
            - gender
            - phone_number
            - languages
            - birth_date
            - emergency_contacts
            - referred_by
            - special_note
          type: object
          properties:
            active:
              type: boolean
              description: whether the patient still visits the clinic.
              example: true
              default: true
    PersonalId:
      required:
      - id
      - type
      type: object
      properties:
        id:
          maxLength: 100
          minLength: 1
          type: string
          description: identifier of the identity card
          example: "3523651512"
        type:
          maxLength: 100
          minLength: 1
          type: string
          description: type of the identity card
          example: ID
      description: identity card
    EmergencyContact:
      required:
      - closeness
      - name
      - phone
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 1
          type: string
          description: name of the contact
          example: Alexa Smith
        closeness:
          maxLength: 100
          minLength: 1
          type: string
          description: closeness of the contact to the patient.
          example: Mother
        phone:
          type: string
          description: "phone number of patient, that includes country code."
          example: +972603159352
...
    IdHolder:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 25
...